Tree + View + Restore #23

+++ src/historyTree.provider.ts
@@ -1,14 +1,20 @@
 import * as vscode from 'vscode';
 
-import {IHistoryFileProperties, HistoryController}  from './history.controller';
-import {IHistorySettings} from './history.settings';
+import { IHistoryFileProperties, HistoryController }  from './history.controller';
+import { IHistorySettings, HistorySettings } from './history.settings';
 
-import path = require('path');
+// import path = require('path');
 
 const enum EHistoryTreeItem {
     None = 0,
     Group,
     File
+}
+
+const enum EHistoryTreeContentKind {
+    Current = 0,
+    All,
+    Search
 }
 
 export default class HistoryTreeProvider implements vscode.TreeDataProvider<HistoryItem>  {
@@ -19,6 +25,7 @@ export default class HistoryTreeProvider implements vscode.TreeDataProvider<Hist
     /* tslint:enable*/
 
     private currentHistoryFile: string;
+    private currentHistoryPath: string;
     private historyFiles: Object; // {yesterday: IHistoryFileProperties[]}
     // save historyItem structure to be able to redraw
     private tree = {};  // {yesterday: {grp: HistoryItem, items: HistoryItem[]}}
@@ -27,12 +34,30 @@ export default class HistoryTreeProvider implements vscode.TreeDataProvider<Hist
     private date;   // calculs result of relative date against now()
     private format; // function to format against locale
 
-    public readonly selectIconPath = {
-        light: path.join(__filename, '..', '..', '..', 'resources', 'images', 'light', 'selection.png'),
-        dark:  path.join(__filename, '..', '..', '..', 'resources', 'images', 'dark', 'selection.png')
-    };
+    public contentKind: EHistoryTreeContentKind = 0;
+    private searchPattern: string;
 
     constructor(private controller: HistoryController) {
+        this.initLocation();
+    }
+
+    initLocation(){
+        vscode.commands.executeCommand('setContext', 'local-history:treeLocation', HistorySettings.getTreeLocation());
+    }
+
+    getSettingsItem(): HistoryItem {
+        // Node only for settings...
+        switch (this.contentKind) {
+            case EHistoryTreeContentKind.All:
+                return new HistoryItem(this, 'Search: all', EHistoryTreeItem.None, null, this.currentHistoryPath);
+                break;
+            case EHistoryTreeContentKind.Current:
+                return new HistoryItem(this, 'Search: current', EHistoryTreeItem.None, null, this.currentHistoryFile);
+                break;
+            case EHistoryTreeContentKind.Search:
+                return new HistoryItem(this, `Search: ${this.searchPattern}`, EHistoryTreeItem.None, null, this.searchPattern);
+                break;
+        }
     }
 
     getTreeItem(element: HistoryItem): vscode.TreeItem {
@@ -47,6 +72,7 @@ export default class HistoryTreeProvider implements vscode.TreeDataProvider<Hist
             if (keys && keys.length) {
                 if (!element) {
                     const items = [];
+                    items.push(this.getSettingsItem());
                     keys.forEach(key => items.push(this.tree[key].grp));
                     return resolve(items);
                 } else if (this.tree[element.label].items) {
@@ -58,6 +84,7 @@ export default class HistoryTreeProvider implements vscode.TreeDataProvider<Hist
             let items: HistoryItem[] = [];
 
             if (!element) { // root
+
                 if (!this.historyFiles) {
 
                     if (!vscode.window.activeTextEditor || !vscode.window.activeTextEditor.document)
@@ -65,19 +92,23 @@ export default class HistoryTreeProvider implements vscode.TreeDataProvider<Hist
 
                     const filename = vscode.window.activeTextEditor.document.uri;
                     const settings = this.controller.getSettings(filename);
+
                     this.loadHistoryFile(filename, settings)
                         .then(() => {
-                            items = this.loadHistoryGroups(this.historyFiles);
+                            items.push(this.getSettingsItem());
+                            items.push(...this.loadHistoryGroups(this.historyFiles));
                             resolve(items);
                         });
                 } else {
-                    items = this.loadHistoryGroups(this.historyFiles);
+                    items.push(this.getSettingsItem());
+                    items.push(...this.loadHistoryGroups(this.historyFiles));
                     resolve(items);
                 }
             } else {
                 if (element.kind === EHistoryTreeItem.Group) {
                     this.historyFiles[element.label].forEach((file) => {
-                        items.push(new HistoryItem(this.format(file), vscode.Uri.file(file.file), element.label));
+                        items.push(new HistoryItem(this, this.format(file), EHistoryTreeItem.File,
+                            vscode.Uri.file(file.file), element.label, true));
                     });
                     this.tree[element.label].items = items;
                 }
@@ -88,29 +119,65 @@ export default class HistoryTreeProvider implements vscode.TreeDataProvider<Hist
 
     private loadHistoryFile(fileName: vscode.Uri, settings: IHistorySettings): Promise<Object> {
         return new Promise((resolve, reject) => {
-            this.controller.findAllHistory(fileName.fsPath, settings, this.noLimit)
-                .then(fileProperties => {
+
+            let pattern;
+            switch (this.contentKind) {
+                case EHistoryTreeContentKind.All:
+                    pattern = '**/*.*';
+                    break;
+                case EHistoryTreeContentKind.Current:
+                    pattern = fileName.fsPath;
+                    break;
+                case EHistoryTreeContentKind.Search:
+                    pattern = this.searchPattern;
+                    break;
+            }
+
+            this.controller.findGlobalHistory(pattern, this.contentKind === EHistoryTreeContentKind.Current, settings, this.noLimit)
+                .then(findFiles => {
                     // Current file
-                    const historyFile = this.controller.decodeFile(fileName.fsPath, settings);
-                    this.currentHistoryFile = historyFile && historyFile.file;
+                    if (this.contentKind === EHistoryTreeContentKind.Current) {
+                        const historyFile = this.controller.decodeFile(fileName.fsPath, settings);
+                        this.currentHistoryFile = historyFile && historyFile.file;
+                    }
+                    this.currentHistoryPath = settings.historyPath;
+
                     // History files
                     this.historyFiles = {};
-                    this.format = (file) => file.date.toLocaleString(settings.dateLocale);
+
+                    this.format = (file) => {
+                        const result = file.date.toLocaleString(settings.dateLocale);
+                        if (this.contentKind !== EHistoryTreeContentKind.Current)
+                            return `${file.name}${file.ext} (${result})`
+                        return result;
+                    };
+
                     let grp = 'new';
-                    const files = fileProperties && fileProperties.history;
+                    const files = findFiles;
                     if (files && files.length)
-                        files.reverse()
-                             .map(file => this.controller.decodeFile(file, settings))
-                             .forEach(file => {
-                                if (grp !== 'Older') {
-                                    grp = this.getRelativeDate(file.date);
-                                    if (!this.historyFiles[grp])
-                                        this.historyFiles[grp] = [file]
-                                    else
+                        files.map(file => this.controller.decodeFile(file, settings))
+                             .sort((f1, f2) => {
+                                if (!f1 || !f2)
+                                    return 0;
+                                if (f1.date > f2.date)
+                                    return -1;
+                                if (f1.date < f2.date)
+                                    return 1;
+                                return f1.name.localeCompare(f2.name);
+                             })
+                             .forEach((file, index) => {
+                                if (file)
+                                    if (grp !== 'Older') {
+                                        grp = this.getRelativeDate(file.date);
+                                        if (!this.historyFiles[grp])
+                                            this.historyFiles[grp] = [file]
+                                        else
+                                            this.historyFiles[grp].push(file);
+                                    } else {
                                         this.historyFiles[grp].push(file);
-                                } else {
-                                    this.historyFiles[grp].push(file);
-                                }
+                                    }
+                                // else
+                                    // this.historyFiles['failed'].push(files[index]);
                             })
                     return resolve(this.historyFiles);
                 })
@@ -123,12 +190,12 @@ export default class HistoryTreeProvider implements vscode.TreeDataProvider<Hist
 
         if (keys && keys.length > 0)
             keys.forEach((key) => {
-                const item =  new HistoryItem(key);
+                const item =  new HistoryItem(this, key, EHistoryTreeItem.Group);
                 this.tree[key] = {grp: item};
                 items.push(item);
             });
         else
-            items.push(new HistoryItem());
+            items.push(new HistoryItem(this, 'No history', EHistoryTreeItem.None));
 
         return items;
     }
@@ -170,18 +237,21 @@ export default class HistoryTreeProvider implements vscode.TreeDataProvider<Hist
             return 'Older'
     }
 
-    private changeItemSelection(select, item) {
-        if (select)
-            item.iconPath = this.selectIconPath
-        else
-            delete item.iconPath;
-    }
+    // private changeItemSelection(select, item) {
+    //     if (select)
+    //          item.iconPath = this.selectIconPath
+    //      else
+    //          delete item.iconPath;
+    // }
 
     private redraw() {
         this._onDidChangeTreeData.fire();
     }
 
     public changeActiveFile() {
+        if (!vscode.window.activeTextEditor)
+            return;
+
         const filename = vscode.window.activeTextEditor.document.uri;
         const settings = this.controller.getSettings(filename);
         const prop = this.controller.decodeFile(filename.fsPath, settings, false);
@@ -194,6 +264,7 @@ export default class HistoryTreeProvider implements vscode.TreeDataProvider<Hist
         delete this.selection;
         this.noLimit = noLimit;
         delete this.currentHistoryFile;
+        delete this.currentHistoryPath;
         delete this.historyFiles;
         delete this.date;
         this._onDidChangeTreeData.fire();
@@ -206,17 +277,47 @@ export default class HistoryTreeProvider implements vscode.TreeDataProvider<Hist
     }
 
     public deleteAll(): void {
-        // delete history for current file
-        this.controller.deleteHistory(this.currentHistoryFile)
-            .then(() => this.refresh());
+        let message;
+        switch (this.contentKind) {
+            case EHistoryTreeContentKind.All:
+                message = `Delete all history - ${this.currentHistoryPath}?`
+                break;
+            case EHistoryTreeContentKind.Current:
+                message = `Delete history for ${this.currentHistoryFile} ?`
+                break;
+            case EHistoryTreeContentKind.Search:
+                message = `Delete history for ${this.searchPattern} ?`
+                break;
+        }
 
-        // const keys = Object.keys(this.historyFiles);
-        // if (keys && keys.length) {
-        //     const items = [];
-        //     keys.forEach(key => items.push(...this.historyFiles[key].map(item => item.file)));
-        //     this.controller.deleteFiles(items)
-        //         .then(() => this.refresh());
-        // }
+        vscode.window.showInformationMessage(message, {modal: true}, {title: 'Yes'}, {title: 'No', isCloseAffordance: true})
+            .then(sel => {
+                if (sel.title === 'Yes') {
+                    switch (this.contentKind) {
+                        case EHistoryTreeContentKind.All:
+                            // Delete all history
+                            this.controller.deleteAll();
+                            break;
+                        case EHistoryTreeContentKind.Current:
+                            // delete history for current file
+                            this.controller.deleteHistory(this.currentHistoryFile)
+                            .then(() => this.refresh());
+                            break;
+                        case EHistoryTreeContentKind.Search:
+                            // Delete visible history files
+                            const keys = Object.keys(this.historyFiles);
+                            if (keys && keys.length) {
+                                const items = [];
+                                keys.forEach(key => items.push(...this.historyFiles[key].map(item => item.file)));
+                                this.controller.deleteFiles(items)
+                                    .then(() => this.refresh());
+                            }
+                            break;
+                    }
+                }
+            },
+                (err => { return; })
+            );
     }
 
     public show(file: vscode.Uri): void {
@@ -240,15 +341,25 @@ export default class HistoryTreeProvider implements vscode.TreeDataProvider<Hist
     }
 
     public compareToCurrent(element: HistoryItem): void {
-        if (element.kind === EHistoryTreeItem.File)
-            this.controller.compare(element.file, vscode.Uri.file(this.currentHistoryFile));
+        if (element.kind === EHistoryTreeItem.File) {
+            let currRange;
+
+            if (vscode.window.activeTextEditor && vscode.window.activeTextEditor.document &&
+                vscode.window.activeTextEditor.document.fileName === this.currentHistoryFile) {
+
+                const currPos = vscode.window.activeTextEditor.selection.active;
+                currRange = new vscode.Range(currPos, currPos);
+            };
+            this.controller.compare(element.file, vscode.Uri.file(this.currentHistoryFile), null, currRange);
+        }
     }
+
     public select(element: HistoryItem): void {
         if (element.kind === EHistoryTreeItem.File) {
             if (this.selection)
                 delete this.selection.iconPath;
             this.selection = element;
-            this.selection.iconPath = this.selectIconPath;
+            // this.selection.iconPath = this.selectIconPath;
             this.tree[element.grp].grp.collapsibleState = vscode.TreeItemCollapsibleState.Expanded;
             this.redraw();
         }
@@ -261,38 +372,79 @@ export default class HistoryTreeProvider implements vscode.TreeDataProvider<Hist
                 vscode.window.showErrorMessage('Select a history files to compare with');
         }
     }
+
+    public restore(element: HistoryItem): void {
+        if (element.kind === EHistoryTreeItem.File) {
+            this.controller.restore(element.file)
+                .then(() => this.refresh())
+                .catch(err => vscode.window.showErrorMessage(`Restore ${element.file.fsPath} failed. Error: ${err}`));
+        }
+    }
+
+    public forCurrentFile(): void{
+        this.contentKind = EHistoryTreeContentKind.Current;
+        this.refresh();
+    }
+    public forAll(): void{
+        this.contentKind = EHistoryTreeContentKind.All;
+        this.refresh();
+    }
+    public forSpecificFile(): void {
+        vscode.window.showInputBox({prompt: 'Specify what to search:', value: '**/*myFile*.*', valueSelection: [4,10]})
+            .then(value => {
+                if (value) {
+                    this.searchPattern = value;
+                    this.contentKind = EHistoryTreeContentKind.Search;
+                    this.refresh();
+                }
+            });
+    }
 }
 
 class HistoryItem extends vscode.TreeItem {
 
     public readonly kind: EHistoryTreeItem;
+    public readonly file: vscode.Uri;
+    public readonly grp: string;
 
-    constructor(label: string = '', public readonly file?: vscode.Uri, public readonly grp: string = '') {
-        super(label != '' ? label : 'No history',
-              file || !label ? vscode.TreeItemCollapsibleState.None : vscode.TreeItemCollapsibleState.Collapsed
-        );
-        this.kind = label ? (file ? EHistoryTreeItem.File : EHistoryTreeItem.Group) : EHistoryTreeItem.None;
+    constructor(provider: HistoryTreeProvider, label: string = '', kind: EHistoryTreeItem, file?: vscode.Uri,
+        grp?: string, showIcon?: boolean) {
+
+        super(label, kind === EHistoryTreeItem.Group ? vscode.TreeItemCollapsibleState.Collapsed : vscode.TreeItemCollapsibleState.None);
+
+        this.kind = kind;
+        this.file = file;
+        this.grp = this.kind !== EHistoryTreeItem.None ? grp : undefined;
+
         switch (this.kind) {
             case EHistoryTreeItem.File:
                 this.contextValue = 'localHistoryItem';
+                this.tooltip = file.fsPath; // TODO remove before .history
+                this.resourceUri = file;
+                if (showIcon)
+                    this.iconPath = false;
                 break;
             case EHistoryTreeItem.Group:
                 this.contextValue = 'localHistoryGrp';
                 break;
             default: // EHistoryTreeItem.None
                 this.contextValue = 'localHistoryNone';
+                this.tooltip = grp;
         }
 
-        // this.command = this.kind === EHistoryTreeItem.File ? {
-        //     command: 'treeLocalHistory.showEntry',
-        //     title: 'Open Local History',
-        //     arguments: [file]
-        // } : undefined;
-
-        this.command = this.kind === EHistoryTreeItem.File ? {
-            command: 'treeLocalHistory.compareToCurrentEntry',
-            title: 'Compare with current version',
-            arguments: [this]
-        } : undefined;
+        // TODO: if current === file
+        if (provider.contentKind === EHistoryTreeContentKind.Current) {
+            this.command = this.kind === EHistoryTreeItem.File ? {
+                command: 'treeLocalHistory.compareToCurrentEntry',
+                title: 'Compare with current version',
+                arguments: [this]
+            } : undefined;
+        } else {
+            this.command = this.kind === EHistoryTreeItem.File ? {
+                command: 'treeLocalHistory.showEntry',
+                title: 'Open Local History',
+                arguments: [file]
+            } : undefined;
+        }
     }
 }
